backend/requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
django-cors-headers==4.3.1 
pymysql==1.1.0 
Pillow==10.0.1 
python-dotenv==1.0.0 
